# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dance Analysis API",
    "description": "Video dance analysis with pose estimation via microservices",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/group-invites/lookup": {
      "get": {
        "tags": [
          "group-invites"
        ],
        "summary": "Lookup Invite",
        "description": "Look up public invite details by token. No authentication required.\n\nUsed by the invite landing page to display group and inviter info\nbefore the user logs in or registers.\n\nReturns 404 for invalid, expired, or revoked tokens.",
        "operationId": "lookup_invite_api_v1_group_invites_lookup_get",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteLookupResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/group-invites/pending": {
      "get": {
        "tags": [
          "group-invites"
        ],
        "summary": "List Pending Invites",
        "description": "List all pending group invites for the authenticated user.\n\nMatches on the user's email address. Returns invites with status=pending\nthat have not yet expired.",
        "operationId": "list_pending_invites_api_v1_group_invites_pending_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GroupInvitePendingResponse"
                  },
                  "type": "array",
                  "title": "Response List Pending Invites Api V1 Group Invites Pending Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/groups/{group_id}/invites": {
      "post": {
        "tags": [
          "group-invites"
        ],
        "summary": "Create Invite",
        "description": "Create a group invite for an email address.\n\n- **email**: Email to invite\n- **role**: Optional role for the invitee\n\nRequires group membership. Does NOT create a membership immediately.",
        "operationId": "create_invite_api_v1_groups__group_id__invites_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Group Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupInviteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupInviteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/group-invites/accept": {
      "post": {
        "tags": [
          "group-invites"
        ],
        "summary": "Accept Invite",
        "description": "Accept a group invite by token.\n\n- **token**: The invite token received\n\nThe accepting user's email must match the invite email (strict).\nReturns 404 for any failure to avoid information leakage.",
        "operationId": "accept_invite_api_v1_group_invites_accept_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInviteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupMembershipResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcceptInviteRequest": {
        "properties": {
          "token": {
            "type": "string",
            "minLength": 1,
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "AcceptInviteRequest",
        "description": "Request body for accepting a group invite by token."
      },
...
      "GroupInviteCreate": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Email address to invite."
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GroupRole"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional role for the invitee (defaults to member)."
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "GroupInviteCreate",
        "description": "Request body for creating a group invite."
      },
      "GroupInvitePendingResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "group_id": {
            "type": "string",
            "format": "uuid",
            "title": "Group Id"
          },
          "group_name": {
            "type": "string",
            "title": "Group Name"
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GroupRole"
              },
              {
                "type": "null"
              }
            ]
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "token",
          "group_id",
          "group_name",
          "role",
          "expires_at",
          "created_at"
        ],
        "title": "GroupInvitePendingResponse",
        "description": "One pending group invite in the authenticated user's invite list."
      },
      "GroupInviteResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "group_id": {
            "type": "string",
            "format": "uuid",
            "title": "Group Id"
          },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GroupRole"
              },
              {
                "type": "null"
              }
            ]
          },
          "token": {
            "type": "string",
            "title": "Token"
          },
          "status": {
            "$ref": "#/components/schemas/GroupInviteStatus"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "accepted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accepted At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "group_id",
          "created_by",
          "email",
          "token",
          "status",
          "expires_at",
          "created_at"
        ],
        "title": "GroupInviteResponse",
        "description": "Response body for a group invite."
      },
      "GroupInviteStatus": {
        "type": "string",
        "enum": [
          "pending",
          "accepted",
          "revoked",
          "expired"
        ],
        "title": "GroupInviteStatus",
        "description": "Status of a group invite."
      },
      "GroupMembershipResponse": {
        "properties": {
          "group_id": {
            "type": "string",
            "format": "uuid",
            "title": "Group Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "role": {
            "$ref": "#/components/schemas/GroupRole"
          },
          "isAdmin": {
            "type": "boolean",
            "title": "Isadmin"
          },
          "status": {
            "$ref": "#/components/schemas/MembershipStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "group_id",
          "user_id",
          "username",
          "role",
          "isAdmin",
          "status",
          "created_at"
        ],
        "title": "GroupMembershipResponse",
        "description": "Response body for a group membership."
      },
...
      "GroupRole": {
        "type": "string",
        "enum": [
          "coach",
          "member"
        ],
        "title": "GroupRole",
        "description": "Role within a group."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InviteLookupResponse": {
        "properties": {
          "group_id": {
            "type": "string",
            "format": "uuid",
            "title": "Group Id"
          },
          "group_name": {
            "type": "string",
            "title": "Group Name"
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GroupRole"
              },
              {
                "type": "null"
              }
            ]
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "group_id",
          "group_name",
          "role",
          "expires_at"
        ],
        "title": "InviteLookupResponse",
        "description": "Public invite details returned before authentication."
      },
...
      "MembershipStatus": {
        "type": "string",
        "enum": [
          "active",
          "invited",
          "removed"
        ],
        "title": "MembershipStatus",
        "description": "Status of a group membership."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}